refactor: group loading components
